change domain type to Mapping (from Dict)
Change domain type from Dict to Mapping.

The `Dict` type is invariant, so if a function takes a `Dict[str, Union[A, B]]`, you cannot pass it a `Dict[str, B]`. This is causing Analytics linting to fail: https://gitlab.com/tumult-labs/analytics/-/jobs/2445736526

The `Mapping` type is covariant, so if a function takes a `Mapping[str, Union[A, B]]`, you _can_ pass a `Dict[str, B]` to the function.

Issue: tumult#1826