Move IMMUTABLE_TYPES.
This is causing an error in Analytics where users will not be able to create sessions unless they have nose.

This is b/c we now use the IMMUTABLE_TYPES data inside of sessions (for copy_if_mutable), but IMMUTABLE_TYPES is in utils/testing which imports nottest from nose.

This can be solved by moving IMMUTABLE_TYPES to type_utils, which avoids Analytics from importing from utils/testing and thus avoids the nose import.