Preliminary version bumping for newer Python version support
As part of tumult#2058, Core needed version updates for certain packages (scipy, numpy, pyarrow) to allow for use with 3.9 and 3.10 - and to allow analytics to use 3.9 and 3.10. With the modifications in this MR, we can support 3.8 and 3.9. 

For 3.10, however, the test suite cannot be run using Nose (which is incompatible with 3.10). Anyone looking to run our tests on 3.10 still can, but they would need to install and use pytest as the test runner instead. Since this is the case, I opted to disallow 3.10 from the versions until core moves away from Nose. This can probably be bumped whenever tumult#2116 is picked up.

For context on the changes related to Numpy: Numpy added typestubs in 1.20.0, so we needed to be more specific with the typing in some cases (mypy threw a tantrum).