Manually load GMP before FLINT/Arb
Installing from source distributions used to not work correctly because pip
would generate wheels that had not been run through delocate or auditwheel, and
thus the GMP shared object couldn't be located when FLINT was loaded. By loading
GMP first, even though it isn't needed directly, this problem is resolved.

This merge request only fixes this problem on Linux-based systems -- a similar 
issue exists on macOS, but it requires a different fix and will be handled
separately.

Also, updates `cibuildwheel` configuration to build wheels for Python 3.9 -- we
claim support for it in our pyproject file, but don't actually build the wheels.