Fix changelog reference
The Python docs have a `Changelog` reference, and if we stop defining one in our
docs all of the changelog links start redirecting to the wrong place with no
warning. This adds a changelog anchor, and also renames it so that an error is
generated if it gets removed again.

Also, cleans up a few minor things in the changelog.