Emit a warning if no Spark session available during cleanup
`utils.cleanup.cleanup()` used to raise an exception if `SparkSession.builder.getOrCreate()` failed, e.g. due to Java not being available. This was true even when run as part of the `atexit` handler, which causes a traceback to be printed when a program exits and may cause confusion. Now, it emits a warning instead if this happens.

tumult#2362