Fix mypy checking
Mypy integration in my editor was broken (maybe this was the case for others as well?). This seems to be related to the fact that running mypy on individual files was broken:
```
> mypy src/tmlt/core/measurements/aggregations.py   
src: error: Source file found twice under different module names: "src.tmlt.core.measurements" and "tmlt.core.measurements"
```
This seems to be due to the fact that mypy knows that `tmlt` is a root of the package, but the combination of settings we use in the mypy configuration also causes the current directory to be used as a root:

```
> cd src/tmlt/core/measurements/
> mypy --config-file ../../../../pyproject.toml aggregations.py
aggregations.py: error: Source file found twice under different module names: "aggregations" and "tmlt.core.measurements.aggregations"
```

This MR allows mypy to be run from the repo base on a specific file.
```
> mypy src/tmlt/core/measurements/aggregations.py
Success: no issues found in 1 source file
```
It does *not* fix the second example. Can I explain why? No. But my editor setup runs mypy from the project base, so this fixes the issue for me. Hopefully this is the way other editors work as well.


Link to Issue: N/A

### For the Assignee:
- [ ] Update the UNRELEASED sections of the changelogs for each project modified, 
        or explain in the MR description why no updates were necessary.

### For the Reviewer:
- [ ] Confirm that an Issue is linked, if applicable.
- [ ] Confirm that the assignee updated relevant changelogs or provided appropriate
    justification as to why no updates were necessary.

This template lives [here](https://gitlab.com/tumult-labs/core/-/blob/dev/.gitlab/merge_request_templates/default_template.md)