Allow N=0 Dataframes in Groupby Statements
This enables groupby aggregations to work with dataframes with a length of zero. This is relevant to the automatic partition selection work, because it is possible that noisy get_groups functions will output an empty dataframe. In that case, we want the program to execute and return an empty dataframe with a matching schema to the expected output schema.

Link to Issue: https://gitlab.com/tumult-labs/tumult/-/issues/2369

### For the Assignee:

- [x] Update the UNRELEASED sections of the changelogs for each project modified, or explain in the MR description why no updates were necessary.

### For the Reviewer:

- [ ] Confirm that an Issue is linked, if applicable.
- [ ] Confirm that the assignee updated relevant changelogs or provided appropriate justification as to why no updates were necessary.

This template lives [here](https://gitlab.com/tumult-labs/core/-/blob/dev/.gitlab/merge_request_templates/default_template.md)