Add support for building ARM-native macOS wheels
ARM-native mac wheels are supported from Python 3.9 onwards; 3.8 is not
officially supported on Apple silicon, only experimentally, so we do not target
it.

Also switches to building the x86_64 macOS wheels using cross-compilation, as
CircleCI is dropping support for Intel mac runners. Due to the limited available
macOS images for Apple silicon, this bumps our minimum macOS version for those
wheels to 12.

In the process of doing the above, changes our CircleCI configuration to only
build macOS wheels on for commits on dev and tags. We may revert this change
later, but since the wheels from feature branches aren't being used it's a waste
of compute to build them.

tumult-labs/tumult#3217