Fix installation instructions
The new installation instructions as part of the changes for https://gitlab.com/tumult-labs/core/-/merge_requests/301 have a bug. The needed path for java is different for arm and x86 architectures, the listed path only applies to x86 architectures. I also switched from messing with the PATH to using the system java wrapper and choosing Java 11 by setting JAVA_HOME, which I think is the nicer/recommended way to do it.