Vendor C dependency sources and include them in source distributions
Source tarballs for GMP, MPFR, FLINT, and Arb are now included in the Core
repository, not downloaded as part of building these dependencies. This should
make building Core from source more resilient to changes or outages on these
projects' download sites/mirrors, and causes this source code to be included in
source distributions of Tumult Core, making it easier to install without
internet access.

As part of this process, there is now a script in ext/ that can be used to
re-download these source tarballs whenever our dependency versions are updated.