Disallow mapping functions that produce input columns when augmenting
Changes the behavior of RowToRowTransformation and RowToRowsTransformation so
that, when augment=True, they raise an exception if trusted_f returns a value
for any of the input columns. Previously such values were ignored and replaced
with the values from the input columns, but this could lead to non-obvious bugs
and didn't have any major benefits.