Improve performance of map output schema checks
Simplifies and speeds up the map output checks introduced in eca25217, as the
slowdown they introduced was fairly significant. This eliminates one redundant
check (though the error message produced when overlapping columns are output
isn't as nice now), and avoids some expensive dictionary copies. The performance
still isn't back to where it was, but it's much closer, especially for tables
that have a large number of columns.