Fix banner config URL in docs version-banner.js
At some point Sphinx removed URL_ROOT from DOCUMENTATION_OPTIONS, which broke
the banner logic. This pulls the same information from a different place in on
the page, and increases the minimum supported version of Sphinx to guarantee
that the new location it pulls from is filled in. It's not included in the
Sphinx changelog, but the relevant change adding support for the new method was
introduced in commit 8e730ae3 in the Sphinx repo.

This change was also made in Analytics in tumult-labs/analytics-upstream!431.