Allow newer sympy versions
This will support our work with Knexus.

As Tom suggested, the problem seems to be that sympy adopted type annotations, but has an assumption system (which provides the `.is_finite` and similar properties) that mypy doesn't understand. Thus, while some of the type issues are fixed by adding annotations, we also need to disable mypy in a few places.