Use new nox_utils from `opendp/tumult-tools` (#2)
* Use new nox_utils from `opendp/tumult-tools`

Switches away from Tumult's original `nox_utils` to a new implementation in
`opendp/tumult-tools`, and makes changes required for that switch. Most notably,
Poetry 2.0 or above is now required, and the pyproject has largely been migrated
to the new format. The noxfile has been updated to use the new `nox_utils` and
to remove various sessions that either we no longer need or are not yet supported.

* Use SessionManager.benchmark to define Nox benchmark sessions

Adds support for running benchmarks in the noxfile using the new
`SessionManager.benchmark` method in nox-utils. Also renames all of the
benchmark scripts to remove the `benchmark_` prefix -- they're already in a
directory called `benchmark`, and it makes some of the scripting in nox-utils a
bit more intuitive.