Switch back to using cibuildwheel for builds, fix wheel tags
Previously, wheels were tagged as `py3-none-any`, but this was incorrect: the
wheels are platform-dependent, just not Python-ABI-dependent. This changes them
to be tagged as `py3-none-<platform>` through some options on Hatchling's custom
build hooks, and re-introduces the custom build function using `cibuildwheel` to
ensure that the build environment is correct.