Permanently disable PLW1641 (eq-without-hash)
This lint flags classes that override __eq__ without overriding __hash__, which
implicitly become unhashable. Core includes a number of classes (various ABCs
and Spark domains, mainly) like this, and it's not a problem that they're
unhashable. We could also resolve this lint by explicitly setting all of their
__hash__ implementations to None, but there's not a clear reason to be concerned
about the implicit behavior for us.