Fix/suppress Ruff's SLF001 (private-member-access)
Mostly just suppressing this, which we did in many places with pylint. Added a
`GroupedDataFrame.dataframe` property, as it has getters for other internal
fields and that field in particular is used a fair amount.

The way suppression comments interact with Black's formatting is a bit funny (it
will sometimes move them onto lines separate from the thing they're suppressing,
and then Ruff will automatically remove them), and produces some less-than-ideal
formatting; switching to Ruff for formatting as well should fix this.