Fix Ruff's PLW0177 (nan-comparison)
This is a slightly odd case in that the rule Ruff is warning about shouldn't
apply here: unlike normal Python, Spark does seem to consider NaN equal to
itself for this purpose. But, the updated version seems clearer to me
anyway. `sf.isnan` has slightly different behavior than comparing to `np.nan`
when applied to nulls (it produces false rather than null), but that shouldn't
matter in this case, as the `when` branch isn't taken when the condition has
either of those values.